app/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bus-kiosk-backend"
version = "1.0.0"
description = "Industrial Django backend for bus kiosk face recognition system"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Imperial Bus Kiosk Team", email = "team@imperial-easypool.com"},
]
maintainers = [
    {name = "Imperial Bus Kiosk Team", email = "team@imperial-easypool.com"},
]
keywords = ["django", "bus", "kiosk", "face-recognition", "industrial"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]

dependencies = [
    # Django Core
    "Django>=5.0,<6.0",
    "djangorestframework>=3.14.0,<4.0",
    "djangorestframework-simplejwt>=5.3.0,<6.0",

    # Database & Infrastructure
    "psycopg2-binary>=2.9.0,<3.0",
    "celery>=5.3.0,<6.0",
    "redis>=5.0.0,<6.0",
    "django-redis>=5.4.0,<6.0",

    # Image Processing (required for ImageField)
    "Pillow>=10.0.0,<11.0",

    # Monitoring & Observability
    "django-prometheus>=2.3.1,<3.0",
    "drf-api-logger>=1.1.12,<2.0",
    "psutil>=5.9.0,<6.0",  # System monitoring

    # Security & Performance
    "django-cors-headers>=4.3.0,<5.0",
    "django-celery-beat>=2.5.0,<3.0",
    "django-celery-results>=2.6.0,<3.0",
    "django-filter>=23.5,<24.0",
    "cryptography>=42.0.0,<43.0",  # For PII encryption
    # Removed django-ratelimit - DRF throttling is sufficient
    "django-csp>=3.8,<4.0",  # Content Security Policy

    # API Enhancement
    "djangorestframework-api-key>=2.3.0,<3.0",  # Better API key management
    # Removed drf-flex-fields - overkill for our use case
    # Removed drf-nested-routers - not needed yet

    # OpenAPI Validation (Industry Standard)
    "openapi-core[django]>=0.18.0,<0.19",  # Production-ready OpenAPI validation
    "openapi-schema-validator>=0.6.3,<0.7",  # OpenAPI schema validation
    "openapi-spec-validator>=0.7.2,<0.8",  # OpenAPI spec validation

    # Utilities
    "python-decouple>=3.8,<4.0",
    "requests>=2.31.0,<3.0",
    "ulid-py>=1.1.0,<2.0",
    "drf-spectacular[sidecar]>=0.27.0,<1.0",  # Self-hosted Swagger UI
    "python-json-logger>=2.0.7,<3.0",
    "django-auditlog>=3.0.0,<4.0",  # Comprehensive audit logging
    "django-model-utils>=4.5.0,<5.0",  # Model utilities
    "argon2-cffi>=23.1.0,<24.0",  # Argon2 password hashing (Fortune 500 security)
]

[tool.setuptools]
# Explicitly define packages to avoid flat-layout issues
packages = ["bus_kiosk_backend", "users", "students", "buses", "events", "kiosks"]

[project.optional-dependencies]
# Development dependencies
dev = [
    # Type checking
    "django-stubs>=5.0.0,<6.0",
    "djangorestframework-stubs>=3.14.0,<4.0",
    "types-PyYAML>=6.0.0,<7.0",
    "types-requests>=2.31.0,<3.0",
    "types-redis>=4.6.0,<5.0",
    "types-psutil>=5.9.0,<6.0",
    "types-django-filter>=25.1,<26.0",

    # Linting & Formatting (Industrial Django Stack)
    "ruff>=0.13.0,<1.0",
    "mypy>=1.18.0,<2.0",
    "black>=25.0.0,<26.0",
    "pre-commit>=3.0.0,<4.0",

    # Development tools
    "django-debug-toolbar>=4.3.0,<5.0",  # Debug toolbar
    "django-extensions>=3.2.3,<4.0",  # Management commands
    "werkzeug>=3.0.0,<4.0",  # Better debugging
]

# Testing dependencies - also need stubs for dev
testing_stubs = [
    "types-factory-boy>=0.4.1,<1.0",  # Type stubs for factory-boy
]


# Testing dependencies
testing = [
    "pytest>=8.0.0,<9.0",
    "pytest-django>=4.5.0,<5.0",
    "pytest-cov>=4.0.0,<5.0",
    # Schemathesis contract testing pinned to 3.x for a stable loader API
    "schemathesis>=3.39.0,<4.0",
    "factory-boy>=3.3.0,<4.0",
    "faker>=24.0.0,<25.0",  # Better test data generation
    "pytest-xdist>=3.5.0,<4.0",  # Parallel test execution
    "pytest-mock>=3.12.0,<4.0",  # Better mocking
    "types-factory-boy>=0.4.1,<1.0",  # Type stubs for factory-boy
]

# Documentation
docs = [
    "sphinx>=7.0.0,<8.0",
    "sphinx-rtd-theme>=1.3.0,<2.0",
]

[project.urls]
Homepage = "https://github.com/easypoolsol-hash/bus_kiosk_imperial"
Repository = "https://github.com/easypoolsol-hash/bus_kiosk_imperial.git"
Issues = "https://github.com/easypoolsol-hash/bus_kiosk_imperial/issues"

[tool.ruff]
# Industrial Django configuration - replaces flake8, isort, and more
target-version = "py310"
line-length = 150

[tool.ruff.lint]
# Core Python rules (comprehensive but fast)
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "B",   # flake8-bugbear (security issues)
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade (syntax upgrades)
    "DJ",  # flake8-django (Django-specific)
    "RUF", # ruff-specific rules
    "I",   # isort (import sorting)
]

ignore = [
    # Django-specific ignores (industrial best practices)
    "DJ001", # null=True on string fields (sometimes needed)
    "DJ008", # __str__ method (not always required)
    "DJ012", # Meta field order (style preference)
    "RUF012", # Mutable class attributes (Django model Meta/indexes/choices)
    # Django patterns that are acceptable
    "BLE001",  # blind except (Django uses this pattern)
    "C901",    # too complex (Django models can be complex)
    "N805",    # self naming (Django conventions)
]

[tool.ruff.lint.per-file-ignores]
# Django-specific file rules
"*/migrations/*.py" = ["E501", "F401"]  # Migrations can be long
"*/settings/*.py" = ["F401"]  # Settings import patterns
"*/tests/*.py" = ["B011", "S101"]  # Test patterns
"**/models.py" = ["RUF012"]  # Django model class attributes (Meta, choices, etc.)
"manage.py" = ["T201"]  # Django management script

[tool.ruff.lint.isort]
# Django-aware import sorting
known-first-party = ["bus_kiosk_backend", "users", "students", "buses", "events", "kiosks"]
force-sort-within-sections = true

[tool.pytest.ini_options]
# Pytest configuration for Django testing
DJANGO_SETTINGS_MODULE = "bus_kiosk_backend.settings"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
    "--tb=short",
    "--cov=bus_kiosk_backend",
    "--cov=users",
    "--cov=students",
    "--cov=buses",
    "--cov=events",
    "--cov=kiosks",
    "--cov-report=xml",
    "--cov-report=term-missing",
    "--cov-fail-under=30",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
# Remove python_classes and python_functions to avoid discovering non-test classes
# Django test discovery will handle proper test classes
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "api: API endpoint tests",
    "slow: Slow running tests",
]

[tool.bandit]
# Security scanning configuration
exclude_dirs = ["tests", "migrations", "node_modules"]
skips = ["B101", "B601"]  # Skip assert checks and shell usage (Django patterns)

[tool.coverage.run]
source = ["bus_kiosk_backend", "users", "students", "buses", "events", "kiosks"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*",
    "manage.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
